fix #680 Logging falls back to Console loggers, can be forced
This commit introduces a Console-backed Logger implementation, which is
now the default fallback for when SLF4J isn't available.

The Loggers factory can be reset or forced to one of the three available
implementations (SLF4J, JDK, Console) via static methods.

The fallback can also be tuned via system property.