fix #770 Add utility method to check if a Throwable is composite
This commit introduces a isMultiple method that check if a Throwable is
a composite created via Exceptions.multiple(...).

The check is made easier internally by the introduction of a package
private CompositeException class, allowing a simple instanceof based
implementation.

Moved and renamed the unit tests for Exceptions to a more consistent
package.